refactor: derive Metadata from its schema
The `z.ZodType<Metadata, unknown>` annotation only caught drift in one
direction: adding an optional field to `Metadata` still typechecked, but
parse stripped the key and the next autosave persisted the stripped
snapshot. Deriving the type from the schema leaves one declaration, so
the two cannot diverge.

Moves `MetadataSchema` next to its sibling schemas in types.ts. Applies
the same fix to `VideoSettings`, whose type alias was a hand-written
duplicate of a schema shape; its only consumer was the old hand-written
`Metadata`, so it goes away.
