+++ lib/project/storeSnapshot.ts
@@ -1,28 +1,6 @@
 import { z } from "zod";
-import { ASPECT_RATIOS } from "@/lib/video/aspectRatio";
-import { TRANSITION_TYPES } from "@/lib/video/transitions";
 import type { ProjectStore } from "./store";
-import {
-	MetadataCharacterSchema,
-	MetadataVoiceSchema,
-	MODES,
-	type Metadata,
-} from "./types";
-
-const VideoSettingsSchema = z.object({
-	transitionType: z.enum(TRANSITION_TYPES).optional(),
-	aspectRatio: z.enum(ASPECT_RATIOS).optional(),
-});
-
-const MetadataSchema: z.ZodType<Metadata, unknown> = z.object({
-	title: z.string().default(""),
-	style: z.string().default(""),
-	narration: MetadataVoiceSchema.default({}),
-	characters: z.record(z.string(), MetadataCharacterSchema).default({}),
-	videoSettings: VideoSettingsSchema.optional(),
-	lastMode: z.enum(MODES).optional(),
-	lastPrompt: z.string().optional(),
-});
+import { MetadataSchema } from "./types";
 
 const ProjectStoreSnapshotSchema = z.object({
 	metadata: z.preprocess((value) => value ?? {}, MetadataSchema),
